Maintain ui element size option added

+++ Diamond Engine/Source Code/CO_Camera.h
@@ -40,8 +40,8 @@ public:
 	float orthoSize;
 	bool cullingState;
 
-	int windowWidth;
-	int windowHeight;
+	float windowWidth;
+	float windowHeight;
 
 //Movement logic
 public:
